requirements.txt
selene==2.0.0rc9
pytest==9.0.2
allure-pytest==2.15.3
selenium==4.35.0
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.13.1